pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ocs-submission"
version = "0.1.0"
description = "OCS Submission Capsule: load FASTQ status, build alignment and post-alignment commands, submit to OCS, audit LIMS, and email summaries."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [{ name = "Allen Institute" }]
dependencies = [
    "pandas>=2.0",
    "psycopg2-binary>=2.9",
    "codeocean>=1.0",
]

[project.scripts]
ocs-submission = "ocs_submission.main:main"

[tool.hatch.build.targets.wheel]
packages = ["ocs_submission"]

[tool.hatch.build.targets.wheel.force-include]
"ocs_submission/config.jsonc" = "ocs_submission/config.jsonc"
"ocs_submission/audit/lims_mtx_ocs.sql" = "ocs_submission/audit/lims_mtx_ocs.sql"
"ocs_submission/audit/lims_rtx_ocs.sql" = "ocs_submission/audit/lims_rtx_ocs.sql"

[tool.hatch.build.targets.sdist]
include = [
    "ocs_submission",
    "README.md",
    "pyproject.toml",
]
